V4.3 Release
The use of unity layers is now changed to using enums on gridGameObject with scripts so there are more options. those being the agent having beter preferences on what terrain to walk.
Many optimizations to the pathfinder and agent scripts.

+++ C# Scripts/Terrain Modifiers/TerrainObject.cs
@@ -13,7 +13,7 @@ public class TerrainObject : MonoBehaviour
     public class CustomSettings
     {
         [Header("Unwalkable = 10, wet_concrete = 20, lower = higher priority")]
-        public float overridePriority;
+        public int overridePriority;
         public int overrideTerrainPenalty;
     }
 }
